+++ cmd/src/repos_delete_metadata.go
@@ -28,17 +28,15 @@ Examples:
 		fmt.Println(usage)
 	}
 	var (
-		repoFlag = flagSet.String("repo", "", `The ID of the repo with the key-value pair metadata to be deleted (required)`)
-		keyFlag  = flagSet.String("key", "", `The name of the  metadata key to be deleted (required)`)
-		apiFlags = api.NewFlags(flagSet)
+		repoFlag     = flagSet.String("repo", "", `The ID of the repo with the key-value pair metadata to be deleted (required if -repo-name is not specified)`)
+		repoNameFlag = flagSet.String("repo-name", "", `The name of the repo to add the key-value pair metadata to (required if -repo is not specified)`)
+		keyFlag      = flagSet.String("key", "", `The name of the  metadata key to be deleted (required)`)
+		apiFlags     = api.NewFlags(flagSet)
 	)
 
 	handler := func(args []string) error {
 		if err := flagSet.Parse(args); err != nil {
 			return err
-		}
-		if *repoFlag == "" {
-			return errors.New("error: repo is required")
 		}
 
 		keyFlag = nil
@@ -53,6 +51,11 @@ Examples:
 		}
 
 		client := cfg.apiClient(apiFlags, flagSet.Output())
+		ctx := context.Background()
+		repoID, err := getRepoIdOrError(ctx, client, repoFlag, repoNameFlag)
+		if err != nil {
+			return err
+		}
 
 		query := `mutation deleteRepoMetadata(
   $repo: ID!,
@@ -67,9 +70,9 @@ Examples:
 }`
 
 		if ok, err := client.NewRequest(query, map[string]interface{}{
-			"repo": *repoFlag,
+			"repo": *repoID,
 			"key":  *keyFlag,
-		}).Do(context.Background(), nil); err != nil || !ok {
+		}).Do(ctx, nil); err != nil || !ok {
 			return err
 		}